Organize wallet module into submodules
