+++ agents/rust/aries-vcx-agent/src/agent/agent_struct.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use aries_vcx::agency_client::agency_client::AgencyClient;
 use aries_vcx::core::profile::profile::Profile;
-use aries_vcx_core::wallet::agency_client_wallet::ToBaseAgencyClientWallet;
+use aries_vcx_core::wallet::indy::agency_client_wallet::ToBaseAgencyClientWallet;
 
 use crate::agent::agent_config::AgentConfig;
