+++ aries_vcx/src/utils/provision.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use agency_client::agency_client::AgencyClient;
 use agency_client::configuration::{AgencyClientConfig, AgentProvisionConfig};
-use aries_vcx_core::wallet::agency_client_wallet::ToBaseAgencyClientWallet;
+use aries_vcx_core::wallet::indy::agency_client_wallet::ToBaseAgencyClientWallet;
 use aries_vcx_core::wallet::base_wallet::BaseWallet;
 
 use crate::errors::error::prelude::*;
