+++ aries_vcx/tests/utils/migration.rs
@@ -11,9 +11,10 @@ use aries_vcx::{
 use aries_vcx_core::{
     indy::wallet::create_and_open_wallet,
     ledger::request_submitter::vdr_ledger::LedgerPoolConfig,
-    wallet::{agency_client_wallet::ToBaseAgencyClientWallet, base_wallet::BaseWallet},
+    wallet::base_wallet::BaseWallet,
     WalletHandle,
 };
+use aries_vcx_core::wallet::indy::agency_client_wallet::ToBaseAgencyClientWallet;
 use aries_vcx_core::wallet::indy::IndySdkWallet;
 use aries_vcx_core::wallet::indy::WalletConfig;
