+++ aries_vcx_core/src/wallet/indy/indy_wallet.rs
@@ -1,7 +1,9 @@
-use async_trait::async_trait;
 use std::collections::HashMap;
-use crate::errors::error::{AriesVcxCoreError, AriesVcxCoreErrorKind, VcxCoreResult};
+
+use async_trait::async_trait;
+
 use crate::{indy, WalletHandle};
+use crate::errors::error::{AriesVcxCoreError, AriesVcxCoreErrorKind, VcxCoreResult};
 use crate::utils::async_fn_iterator::AsyncFnIterator;
 use crate::wallet::base_wallet::BaseWallet;
 use crate::wallet::indy::{IndySdkWallet, IndyWalletRecordIterator, internal, WalletRecord};
@@ -90,7 +92,7 @@ impl BaseWallet for IndySdkWallet {
         xtype: &str,
         query: &str,
         options: &str,
-    ) -> VcxCoreResult<Box<dyn AsyncFnIterator<Item = VcxCoreResult<String>>>> {
+    ) -> VcxCoreResult<Box<dyn AsyncFnIterator<Item=VcxCoreResult<String>>>> {
         let search = internal::open_search_wallet(self.wallet_handle, xtype, query, options).await?;
         let iter = IndyWalletRecordIterator::new(self.wallet_handle, search);
