--- aries_vcx_core/src/wallet/indy_wallet.rs
@@ -1,1 +1,0 @@
-
