+++ aries_vcx_core/src/wallet/mod.rs
@@ -1,6 +1,3 @@
-pub mod agency_client_wallet;
 pub mod base_wallet;
 #[cfg(feature = "vdrtools")]
-pub mod indy_wallet;
-pub mod internal;
 pub mod indy;
